v0.0.9a
Refactored the Sound Emitter to the base Tapestry_Activatable class.

Added sound inputs for Props when pushing. Added sound input for pickup and collision sound for Items, but haven't hooked them into the inspector yet.

Moved the health controls for Props into the Destructable tab. Exposed interactable and display name parameters in its place at the top of the inspector.